requirements.txt
PyQt5==5.15.9
psycopg2-binary==2.9.9 